VSCode: Separate "main" functions from LanguageClient code
commit-id:13cfed95
